迁移about页面至ui/src
