+++ apps/workshop/src/stories/pages/pages.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import HomeV2311 from './home-v2311'
 import PdpIntroV2311 from './pdp-intro-v2311'
 import PdpUserStoryV2311 from './pdp-user-story-v2311'
-import About from './about'
+import About from 'ui/src/about'
 
 const meta: Meta = {
   title: 'pages/pathes',
